# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Bal Suraksha API
  version: "0.1.0"
  description: |
    Private child-safety support platform. All endpoints enforce organization-scoped
    authorization. Enums are fixed; unknown values are rejected server-side.
  contact:
    name: Bal Suraksha engineering
servers:
  - url: http://localhost:8080/api/v1
    description: Local development

# ── Security schemes ────────────────────────────────────────────────────────
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      description: Short-lived JWT issued on private session creation.
    staffToken:
      type: http
      scheme: bearer
      description: JWT issued on staff login (local stub for demo; OIDC in production).

  # ── Shared error envelope ─────────────────────────────────────────────────
  schemas:
    ErrorEnvelope:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable error message. Never contains private case content.
        code:
          type: string
          description: Machine-readable error code.
        request_id:
          type: string

    # ── Enums ─────────────────────────────────────────────────────────────
    CaseStatus:
      type: string
      enum:
        - RECEIVED
        - TRIAGE
        - ASSIGNED
        - ACCEPTED
        - IN_PROGRESS
        - FOLLOW_UP
        - CLOSED
...
    CaseCreateRequest:
      type: object
      required: [account_text]
      properties:
        account_text:
          type: string
          maxLength: 10000
        safe_contact_preference:
          $ref: "#/components/schemas/SafeContactPreference"
        conflict_flag:
          type: string
          nullable: true
          enum: [school_implicated, caregiver_implicated, staff_implicated]
        optional_safe_contact:
          type: string
          nullable: true
        route_type:
          type: string
          enum: [ask_for_help, worried_about_someone, missing_child]
    SafeContactPreference:
      type: object
      properties:
        time_zone:
          type: string
          enum: [Asia/Kolkata, UTC]
          default: Asia/Kolkata
        safe_hours_start:
          type: string
          nullable: true
        safe_hours_end:
          type: string
          nullable: true
        preferred_channel:
          type: string
          enum: [message_in_app, no_contact]
          nullable: true
    CaseCreateResponse:
      type: object
      required: [case_id, receipt_id, status]
      properties:
        case_id:
          type: string
          format: uuid
        receipt_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/CaseStatus"
        received_at:
          type: string
          format: date-time
    CaseSafeView:
      type: object
      description: Child-visible view. Never includes staff notes or other reporter data.
      required: [case_id, status, last_updated_at]
      properties:
        case_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/CaseStatus"
        last_updated_at:
          type: string
          format: date-time
        safe_progress_message:
          type: string
          description: Plain-language status update safe for the child to read.
          nullable: true
        messages:
          type: array
          items:
            $ref: "#/components/schemas/SafeMessageView"
    SafeMessageView:
      type: object
      required: [id, body, sent_at, sender_role]
      properties:
        id:
          type: string
          format: uuid
        body:
          type: string
        sent_at:
          type: string
          format: date-time
        sender_role:
          type: string
          enum: [reporter, responder]
    CaseStaffView:
      type: object
      description: Responder/staff view. Restricted to assigned organization.
      required: [case_id, status, created_at, route_type]
      properties:
        case_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/CaseStatus"
        route_type:
          type: string
        conflict_flag:
          type: string
          nullable: true
        account_text:
          type: string
        safe_contact_preference:
          $ref: "#/components/schemas/SafeContactPreference"
          nullable: true
        created_at:
          type: string
          format: date-time
        assigned_to:
          type: string
          format: uuid
          nullable: true
        assignment_accepted_at:
          type: string
          format: date-time
          nullable: true
        version:
          type: integer

    # ── Assignments ───────────────────────────────────────────────────────
...
paths:
...
  /cases:
    post:
      operationId: createCase
      summary: Submit a support request
      tags: [cases]
      security:
        - sessionToken: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
          description: Client-generated UUID. Retry with same key returns original receipt.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseCreateRequest"
      responses:
        "201":
          description: Case created and durably received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseCreateResponse"
        "200":
          description: Existing case returned (idempotent retry)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseCreateResponse"
        "409":
          description: Same idempotency key with different payload rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /cases/{case_id}/safe-view:
    get:
      operationId: getCaseSafeView
      summary: Child-visible case progress (restricted fields only)
      tags: [cases]
      security:
        - sessionToken: []
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Safe case view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseSafeView"
        "403":
          description: Not authorized for this case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /staff/cases:
    get:
      operationId: listStaffCases
      summary: List cases for the authenticated staff member's organization
      tags: [staff-cases]
      security:
        - staffToken: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/CaseStatus"
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of cases (org-scoped)
          content:
            application/json:
              schema:
                type: object
                properties:
                  cases:
                    type: array
                    items:
                      $ref: "#/components/schemas/CaseStaffView"
                  total:
                    type: integer
  /staff/cases/{case_id}:
    get:
      operationId: getStaffCase
      summary: Get case detail (staff, org-scoped)
      tags: [staff-cases]
      security:
        - staffToken: []
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Case detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStaffView"
        "403":
          description: Not authorized for this case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # Assignments
...